requirements.txt
gradio==4.7.1
gradio_client==0.7.0
huggingface-hub==0.19.4
langchain==0.0.344
matplotlib==3.8.2
numpy==1.26.2
openai==0.28.0
opencv-python==4.8.1.78
pandas==2.1.3
pdf2image==1.16.3
Pillow==10.1.0
pinecone-client==2.2.4
pypdf==3.17.1
regex==2023.10.3
scikit-learn==1.3.2
scipy==1.11.4
sentence-transformers==2.2.2
sentencepiece==0.1.99
streamlit==1.29.0
tokenizers==0.15.0
torch==2.1.1
torchvision==0.16.1
tqdm==4.66.1
transformers==4.35.2
streamlit_chat

